requirements.txt
streamlit
pandas
numpy
scikit-learn
plotly
mlxtend
python-pptx
requests
faker
